update to typescript & antd4.0
