+++ package.json
@@ -22,8 +22,9 @@
     "assets"
   ],
   "dependencies": {
+    "@ant-design/icons": "^4.0.3",
     "react-resizable": "^1.10.1",
-    "antd": "^3.17.0"
+    "antd": "^4.0.3"
   },
   "peerDependencies": {
     "react": "^16.12.0",
