+++ src/index.less
@@ -23,7 +23,7 @@
       > div {
         line-height: 24px;
       }
-      > i > svg {
+      > span > svg {
         margin-top: 6px;
         margin-right: 16px;
         cursor: pointer;
@@ -38,7 +38,7 @@
     .antETableToolbarRight {
       width: 20%;
       text-align: right;
-      > i > svg {
+      > span > svg {
         margin-top: 6px;
         margin-right: 16px;
         cursor: pointer;
